KFSPTS-21519 making adding and editing a PAAT document through batch job transactional
