+++ src/main/java/edu/cornell/kfs/pdp/batch/service/PayeeACHAccountDocumentService.java
@@ -1,0 +1,14 @@
+package edu.cornell.kfs.pdp.batch.service;
+
+import org.kuali.kfs.pdp.businessobject.PayeeACHAccount;
+import org.kuali.rice.kim.api.identity.Person;
+
+import edu.cornell.kfs.pdp.businessobject.PayeeACHAccountExtractDetail;
+
+public interface PayeeACHAccountDocumentService {
+    
+    String addACHAccount(Person payee, PayeeACHAccountExtractDetail achDetail, String payeeType);
+    String updateACHAccountIfNecessary(Person payee, PayeeACHAccountExtractDetail achDetail, PayeeACHAccount achAccount);
+    String getDirectDepositTransactionType();
+
+}
